chore: Tolerate other analytics events in flashbar analytics tests
